requirements.txt
asgiref==3.6.0
attrs==22.2.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
cryptography==40.0.1
defusedxml==0.7.1
Django==3.2.16
django-filter==23.1
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
djoser==2.1.0
flake8==6.0.0
idna==3.4
iniconfig==2.0.0
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.2
mccabe==0.7.0
oauthlib==3.2.2
packaging==23.0
Pillow==9.3.0
pluggy==0.13.1
py==1.11.0
pycodestyle==2.10.0
pycparser==2.21
pyflakes==3.0.1
PyJWT==2.1.0
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
python3-openid==3.2.0
pytz==2023.2
requests==2.26.0
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.4.0
sqlparse==0.4.3
toml==0.10.2
uritemplate==4.1.1
urllib3==1.26.15
